update naming for test order
